add macro support for custom reader fn; begin implementing player position-based tile section syncing

+++ src/network/utils.rs
@@ -4,26 +4,26 @@ pub fn flags(a: bool, b: bool, c: bool, d: bool, e: bool, f: bool, g: bool, h: b
 	a as u8 | (b as u8) << 1 | (c as u8) << 2 | (d as u8) << 3 | (e as u8) << 4 | (f as u8) << 5 | (g as u8) << 6 | (h as u8) << 7
 }
 
-pub fn get_section_x(x: i32) -> i32 {
+pub fn get_section_x(x: usize) -> usize {
 	x / 200
 }
 
-pub fn get_section_y(y: i32) -> i32 {
+pub fn get_section_y(y: usize) -> usize {
 	y / 150
 }
 
-pub fn get_tile_x_start(sec_x: i32) -> i32 {
+pub fn get_tile_x_start(sec_x: usize) -> usize {
 	sec_x * 200
 }
 
-pub fn get_tile_x_end(sec_x: i32) -> i32 {
+pub fn get_tile_x_end(sec_x: usize) -> usize {
 	(sec_x + 1) * 200
 }
 
-pub fn get_tile_y_start(sec_y: i32) -> i32 {
+pub fn get_tile_y_start(sec_y: usize) -> usize {
 	sec_y * 150
 }
 
-pub fn get_tile_y_end(sec_y: i32) -> i32 {
+pub fn get_tile_y_end(sec_y: usize) -> usize {
 	(sec_y + 1) * 150
 }
